Improve filtering on retry attempts when loading Talkdesk data
In this script if an exception is raised during import, the retry mechanism
will redownload the whole day where a failure happened. This is not scalable
as some (most?) days contain thousands of calls.

This MR adds a mechanism to check which days have missing metadata in Snowflake
and filter in order to only include those on retry attempts.